+++ src/transaction.rs
@@ -145,6 +145,14 @@ pub enum PostingPrice<D> {
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct Tag(Arc<str>);
 
+impl Tag {
+    /// Returns underlying string representation
+    #[must_use]
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
 impl Display for Tag {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         Display::fmt(&self.0, f)
@@ -182,6 +190,14 @@ impl Borrow<str> for Tag {
 /// ```
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
 pub struct Link(Arc<str>);
+
+impl Link {
+    /// Returns underlying string representation
+    #[must_use]
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
 
 impl Display for Link {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {